stock_service/requirements/base.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
idna==2.10
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
psycopg2==2.9.2
kombu==5.2.2
Django==3.1.7
django-allauth==0.43.0
# Django REST Framework
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
dj-rest-auth==2.2.3
# Swagger
drf-yasg==1.20.0
